feat: linking track section front and back end
