+++ .prettierrc
@@ -1,10 +1,9 @@
 {
     "semi": true,
     "singleQuote": true,
-    "tabWidth": 4,
     "trailingComma": "all",
-    "printWidth": 80,
-    "bracketSpacing": true,
     "arrowParens": "avoid",
-    "useTabs": false
+    "useTabs": false,
+    "tabWidth": 4,
+    "printWidth": 999
 }
