+++ src/components/Configuracoes/Sections/CriterionSections.ts
@@ -1,0 +1,3 @@
+export type SectionType = 'Pilares' | 'Trilhas' | 'Ciclos';
+
+export const criterionSections: SectionType[] = ['Pilares', 'Trilhas', 'Ciclos'];
