--- src/components/Criterios/Sections/CriterionSections.ts
@@ -1,7 +1,0 @@
-export type SectionType = 'Pilares' | 'Trilhas' | 'Ciclos';
-
-export const criterionSections: SectionType[] = [
-    'Pilares',
-    'Trilhas',
-    'Ciclos',
-];
