+++ src/router/index.tsx
@@ -9,7 +9,7 @@ import { MultiRoleRoute } from './MultiRoleRoute';
 import { UserRoleEnum } from '../types/auth';
 import { Dashboard } from '../pages/Dashboard';
 import { Colaboradores } from '../pages/Colaboradores';
-import { Criterios } from '../pages/RH/Criterios';
+import { Configuracoes } from '../pages/RH/Configuracoes';
 import { ImportarHistoricos } from '../pages/RH/ImportarHistoricos';
 import { Avaliacao } from '../pages/Colaborador/Avaliacao';
 import { Evolucao } from '../pages/Colaborador/Evolucao';
@@ -33,25 +33,13 @@ export function Router() {
     return (
         <Routes>
             {/* Login: se já estiver autenticado, manda direto pro dashboard */}
-            <Route
-                path="/login"
-                element={
-                    isAuthenticated ? (
-                        <Navigate to="/dashboard" replace />
-                    ) : (
-                        <LoginPage />
-                    )
-                }
-            />
+            <Route path="/login" element={isAuthenticated ? <Navigate to="/dashboard" replace /> : <LoginPage />} />
 
             {/* Roteamento protegido */}
             <Route element={<ProtectedRoute />}>
                 <Route element={<DefaultLayout />}>
                     {/* redireciona / para /dashboard */}
-                    <Route
-                        index
-                        element={<Navigate to="dashboard" replace />}
-                    />
+                    <Route index element={<Navigate to="dashboard" replace />} />
 
                     <Route path="dashboard" element={<Dashboard />} />
                     <Route path="avaliacao" element={<Avaliacao />} />
@@ -68,23 +56,17 @@ export function Router() {
                     <Route
                         path="colaboradores"
                         element={
-                            <RoleRoute
-                                requiredRoles={[
-                                    UserRoleEnum.RH,
-                                    UserRoleEnum.MENTOR,
-                                    UserRoleEnum.LEADER,
-                                ]}
-                            >
+                            <RoleRoute requiredRoles={[UserRoleEnum.RH, UserRoleEnum.MENTOR, UserRoleEnum.LEADER]}>
                                 <Colaboradores />
                             </RoleRoute>
                         }
                     />
 
                     <Route
-                        path="criterios"
+                        path="configuracoes"
                         element={
                             <RoleRoute requiredRoles={[UserRoleEnum.RH]}>
-                                <Criterios />
+                                <Configuracoes />
                             </RoleRoute>
                         }
                     />
@@ -103,13 +85,8 @@ export function Router() {
                         element={
                             <RoleRoute requiredRoles={[UserRoleEnum.RH]}>
                                 <div className="p-6">
-                                    <h1 className="text-2xl font-bold">
-                                        Painel de Administração
-                                    </h1>
-                                    <p>
-                                        Esta página só é acessível para RH,
-                                        Comitê, Admin e Desenvolvedor
-                                    </p>
+                                    <h1 className="text-2xl font-bold">Painel de Administração</h1>
+                                    <p>Esta página só é acessível para RH, Comitê, Admin e Desenvolvedor</p>
                                 </div>
                             </RoleRoute>
                         }
@@ -118,18 +95,10 @@ export function Router() {
                     <Route
                         path="dev"
                         element={
-                            <RoleRoute
-                                requiredRoles={[UserRoleEnum.DEVELOPER]}
-                                redirectTo="/dashboard"
-                            >
+                            <RoleRoute requiredRoles={[UserRoleEnum.DEVELOPER]} redirectTo="/dashboard">
                                 <div className="p-6">
-                                    <h1 className="text-2xl font-bold">
-                                        Ferramentas de Desenvolvimento
-                                    </h1>
-                                    <p>
-                                        Esta página só é acessível para
-                                        desenvolvedores
-                                    </p>
+                                    <h1 className="text-2xl font-bold">Ferramentas de Desenvolvimento</h1>
+                                    <p>Esta página só é acessível para desenvolvedores</p>
                                 </div>
                             </RoleRoute>
                         }
@@ -138,22 +107,10 @@ export function Router() {
                     <Route
                         path="mentoria"
                         element={
-                            <MultiRoleRoute
-                                allowedRoles={[
-                                    UserRoleEnum.MENTOR,
-                                    UserRoleEnum.LEADER,
-                                    UserRoleEnum.MANAGER,
-                                ]}
-                                redirectTo="/dashboard"
-                            >
+                            <MultiRoleRoute allowedRoles={[UserRoleEnum.MENTOR, UserRoleEnum.LEADER, UserRoleEnum.MANAGER]} redirectTo="/dashboard">
                                 <div className="p-6">
-                                    <h1 className="text-2xl font-bold">
-                                        Mentoria
-                                    </h1>
-                                    <p>
-                                        Esta página é acessível para mentores,
-                                        líderes e gestores
-                                    </p>
+                                    <h1 className="text-2xl font-bold">Mentoria</h1>
+                                    <p>Esta página é acessível para mentores, líderes e gestores</p>
                                 </div>
                             </MultiRoleRoute>
                         }
@@ -162,15 +119,7 @@ export function Router() {
             </Route>
 
             {/* Catch-all: qualquer rota não encontrada redireciona para login ou dashboard */}
-            <Route
-                path="*"
-                element={
-                    <Navigate
-                        to={isAuthenticated ? '/dashboard' : '/login'}
-                        replace
-                    />
-                }
-            />
+            <Route path="*" element={<Navigate to={isAuthenticated ? '/dashboard' : '/login'} replace />} />
         </Routes>
     );
 }
